Refector game loop to new crate
