+++ sdl-frontend/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "sdl-frontend"
+version = "0.1.0"
+authors = ["Henry Sloan <henryksloan@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+nes = { path = "../nes" }
+sdl2 = "0.34.3"
