+++ windows-ui/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sdl-ui"
+name = "windows-ui"
 version = "0.1.0"
 authors = ["Henry Sloan <henryksloan@gmail.com>"]
 edition = "2018"
@@ -8,6 +8,7 @@ edition = "2018"
 nes = { path = "../nes" }
 cpu = { path = "../cpu" }
 memory = { path = "../memory" }
+sdl-frontend = { path = "../sdl-frontend" }
 sdl2 = "0.34.3"
 sdl2-sys = "0.34.3"
 rand = "0.8.3"
